% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armalda}
\alias{armalda}
\title{This function implements the Linear Discriminant Analysis (LDA) algorithm to reduce the dimensionality of data to the specified number of components.
The input parameters include the data matrix X, the class label vector y, and the number of components num_components.}
\usage{
armalda(X, y, num_components)
}
\arguments{
\item{X}{Data matrix, where each row represents a sample and each column represents a feature.}

\item{y}{Class label vector indicating the class membership of each sample.}

\item{num_components}{Number of dimensions for LDA dimensionality reduction.}
}
\value{
The reduced-dimensional data matrix, where each row represents a sample and each column represents a reduced feature.
}
\description{
This function implements the Linear Discriminant Analysis (LDA) algorithm to reduce the dimensionality of data to the specified number of components.
The input parameters include the data matrix X, the class label vector y, and the number of components num_components.
}
