% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armalle}
\alias{armalle}
\title{Function to compute the Locally Linear Embedding (LLE) algorithm
This function takes the input data matrix, along with the parameters k and d, and returns the LLE embedding.}
\usage{
armalle(data, k, d)
}
\arguments{
\item{data}{Data matrix, where each row represents a sample and each column represents a feature.}

\item{k}{Number of nearest neighbors to consider for each sample.}

\item{d}{Number of dimensions for the LLE embedding.}
}
\value{
The LLE embedding as a numeric matrix, where each row represents a sample and each column represents a reduced feature.
}
\description{
Function to compute the Locally Linear Embedding (LLE) algorithm
This function takes the input data matrix, along with the parameters k and d, and returns the LLE embedding.
}
