% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armaPCA}
\alias{armaPCA}
\title{Perform Principal Component Analysis (PCA) on a dataset.}
\usage{
armaPCA(data, num_components = -1L, variance_percentage = -1)
}
\arguments{
\item{data}{A numeric matrix containing the input data.}

\item{num_components}{The number of principal components to retain (default: -1).
If set to -1, all components will be retained.}

\item{variance_percentage}{The percentage of variance to be explained by the retained components (default: -1).
If set to -1, all components will be retained.}
}
\value{
A numeric matrix containing the projected data in the principal component space.
}
\description{
This function takes a numeric matrix as input and performs PCA on the dataset.
The function allows the user to specify the number of principal components to retain
or the percentage of variance to be explained by the retained components.
}
