% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armaTsne}
\alias{armaTsne}
\title{R 包接口函数
This function serves as the interface between R and the t-SNE algorithm implemented in C++.
It takes the input data matrix, along with optional parameters, and returns the t-SNE embedding.}
\usage{
armaTsne(
  data,
  num_dims = 2L,
  num_iters = 1000L,
  learning_rate = 200,
  perplexity = 30
)
}
\arguments{
\item{data}{Data matrix, where each row represents a sample and each column represents a feature.}

\item{num_dims}{Number of dimensions for the t-SNE embedding(default: 2).}

\item{num_iters}{Number of iterations for the t-SNE algorithm(default: 1000).}

\item{learning_rate}{Learning rate for the t-SNE algorithm(default: 200.0).}

\item{perplexity}{Perplexity parameter for the t-SNE algorithm(default: 30.0).}
}
\value{
The t-SNE embedding as a numeric matrix, where each row represents a sample and each column represents a reduced feature.
}
\description{
R 包接口函数
This function serves as the interface between R and the t-SNE algorithm implemented in C++.
It takes the input data matrix, along with optional parameters, and returns the t-SNE embedding.
}
